project/app/requirements.txt
# Web Framework and UI
fastapi
uvicorn[standard]
pydantic
nicegui

# MCP (Model Context Protocol)
mcp
fastmcp

# Google AI and APIs
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# PDF Processing
pymupdf  # Provides fitz module
pypdf
# Document Processing & RAG
chromadb
sentence-transformers

# Email Service (Bypasses SMTP block on Render)
resend

# Utilities
python-dotenv
